Write a tiny script to filter projects with multiple remotes
This is useful to identify all the Emacs packages I'm contributing in.
The reason I needed to do this was that I deleted my .emacs.d folder
and re-installed from scratch after a long time today. I wanted to be
sure that I don't lose any local modifications, so I used this script
to double-check all the places where I have a personal fork.