Make changes to try and fix the tmux problem
`tmux` was using `tmux-256color` instead of `xterm-256color` as the
`$TERM`, and this was causing the terminal to do weird things. These
settings did not change anything. The correct fix was to reinstall the
`tmux-256color` code using `tic -x tmux256-color` on the new file in
this commit.

This was taken from
https://gist.github.com/nicm/ea9cf3c93f22e0246ec858122d9abea1,
committing here for easy reference if the problem re-occurs in the
future.